pyproject.toml
[tool.poetry]
name = "legal-tech-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
streamlit = "*"
openai = "^1.0.0"
pydantic = "^2.0.0"
lancedb = "*"
sentence-transformers = "^2.2.2"
python-dotenv = "*"
instructor = "*"
tiktoken = "*"
eyecite = "*"
citeurl = "*"
tantivy = "^0.21.0"
requests = "^2.31.0"
black = {extras = ["jupyter"], version = "^23.3.0"}
backoff = "^2.2.1"
html2text = "^2020.1.16"
datamodel-code-generator = "^0.25.3"
devtools = "^0.12.2"
spacy = "^3.7.2"
thefuzz = ">=0.20.0,<0.21.0"
suffix-trees = "^0.3.0"
tabulate = "^0.9.0"
termcolor = "^2.4.0"
exa-py = "^1.0.7"
qdrant-client = "^1.7.2"
momento = "^1.17.0"
redis = "^5.0.1"
fakeredis = "^2.21.0"
pyparsing = "^3.1.1"
colorlog = "^6.8.2"
bs4 = "^0.0.2"
pygithub = "^1.58.1"
python-docx = "^1.1.0"
lxml = "^4.9.3"
unstructured = {extras = ["docx", "pptx", "pdf"], version = ">=0.10.16,<0.10.18"}
haystack = "^0.42"
fire = "^0.5.0"
trafilatura = "^1.5.0"
pypdf = "^3.12.2"
rank-bm25 = "^0.2.2"
pdfplumber = "^0.10.2"
pymupdf = "^1.23.3"
faker = "^22.6.0"
datasketch = "^1.6.4"
metaphor-python = "^0.1.23"
google-api-python-client = "^2.95.0"
pytest = "^8.0.0"
pydantic-settings = "^2.1.0"
graphviz = "^0.20.1"
chromadb = "^0.4.22"
diskcache = "^5.6.3"
md2pdf = "^1.0.1"
aiofiles = "^23.2.1"
markdown = "^3.5.2"
duckdb = "^0.9.2"
justopinion = "^0.2.5"
dspy-ai = {extras = ["chromadb"], version = "^2.1.10"}
pytest-asyncio = "^0.23.5"
geopandas = "^0.14.3"
us = "^3.1.1"
seaborn = "^0.13.2"
pyquery = "^2.0.0"
requests-cache = "^1.2.0"
openpyxl = "^3.1.2"
faiss-cpu = "^1.8.0"
pdfkit = "^1.0.0"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = [
    "tests",
]
pythonpath = [
    "src",
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
notebook = "^7.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
